requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
aiosqlite>=0.19.0
asyncpg>=0.29.0
httpx>=0.26.0
prometheus-client>=0.19.0
jinja2>=3.1.0
python-multipart>=0.0.6
websockets>=12.0
apscheduler>=3.10.0
bcrypt>=4.1.0
pyjwt>=2.8.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
slowapi>=0.1.9
psycopg2-binary>=2.9.0
